+++ include/battle_ai_script_commands.h
@@ -10,7 +10,7 @@ void BattleAI_HandleItemUseBeforeAISetup(u8 defaultScoreMoves);
 void BattleAI_SetupAIData(u8 defaultScoreMoves);
 u8 BattleAI_ChooseMoveOrAction(void);
 void ClearBattlerMoveHistory(u8 battlerId);
-void RecordAbilityBattle(u8 battlerId, u8 abilityId);
+void RecordAbilityBattle(u8 battlerId, u16 abilityId);
 void ClearBattlerAbilityHistory(u8 battlerId);
 void RecordItemEffectBattle(u8 battlerId, u8 itemEffect);
 void ClearBattlerItemEffectHistory(u8 battlerId);
